test - update
Moving the NetworkDeltaPositionTests into its own file.
